test: split test-utils to prevent interaction
