+++ src/components/PositionListItem/PositionListItem.test.tsx
@@ -5,7 +5,8 @@ import { USDC_MAINNET } from 'constants/tokens'
 import { useToken } from 'hooks/Tokens'
 import { usePool } from 'hooks/usePools'
 import { PoolState } from 'hooks/usePools'
-import { mocked, render } from 'test-utils'
+import { mocked } from 'test-utils/mocked'
+import { render } from 'test-utils/render'
 import { unwrappedToken } from 'utils/unwrappedToken'
 
 import PositionListItem from '.'
