+++ src/components/SearchModal/CurrencyList/index.test.tsx
@@ -1,7 +1,7 @@
 import { Currency, CurrencyAmount as mockCurrencyAmount, Token as mockToken } from '@uniswap/sdk-core'
 import { DAI, USDC_MAINNET, WBTC } from 'constants/tokens'
 import * as mockJSBI from 'jsbi'
-import { render } from 'test-utils'
+import { render } from 'test-utils/render'
 
 import CurrencyList from '.'
