+++ src/components/WalletDropdown/MiniPortfolio/Pools/index.test.tsx
@@ -2,7 +2,8 @@ import { BigNumber } from '@ethersproject/bignumber'
 import { SupportedChainId, WETH9 } from '@uniswap/sdk-core'
 import { FeeAmount, Pool, Position } from '@uniswap/v3-sdk'
 import { USDC_MAINNET } from 'constants/tokens'
-import { mocked, render } from 'test-utils'
+import { mocked } from 'test-utils/mocked'
+import { render } from 'test-utils/render'
 
 import Pools from '.'
 import useMultiChainPositions from './useMultiChainPositions'
