+++ src/components/WalletDropdown/MiniPortfolio/PortfolioLogo.test.tsx
@@ -1,7 +1,7 @@
 import { SupportedChainId } from '@uniswap/sdk-core'
 import { DAI_ARBITRUM } from '@uniswap/smart-order-router'
 import { DAI, USDC_ARBITRUM, USDC_MAINNET } from 'constants/tokens'
-import { render } from 'test-utils'
+import { render } from 'test-utils/render'
 
 import { PortfolioLogo } from './PortfolioLogo'
