+++ src/nft/components/bag/bag.test.tsx
@@ -1,4 +1,4 @@
-import { render } from 'test-utils'
+import { render } from 'test-utils/render'
 
 import Bag from './Bag'
