+++ src/pages/Pool/CTACards.test.tsx
@@ -1,5 +1,5 @@
 import * as useV3Positions from 'hooks/useV3Positions'
-import { render, screen } from 'test-utils'
+import { render, screen } from 'test-utils/render'
 
 import CTACards from './CTACards'
